+++ .github/workflows/release-notes.md
@@ -5,10 +5,12 @@
 
 ### New features
 - Added following/followers count on "Profile"
+- Added user's following/followers list accessible from "Profile" or user's page
 - Added back button on all pages
 - Added "Follow" button on users list
 - Cleaned all users lists by removing timestamp
 - Added total number of users in "Users" page
+
 
 ### Fixed bugs
 - Address visualization on "User's details" dialog, when opened from "Notifications" page